Refactored AmazonTransport.request() and index.js
Refactored for code readability and one class per file
