+++ tests/AmazonTransport.test.js
@@ -3,31 +3,30 @@
 const { Transport } = require('@elastic/elasticsearch')
 const assert = require('assert')
 
-const { AmazonTransport } = require('../src')
+const AmazonTransport = require('../src/AmazonTransport')
 
 describe('AmazonTransport', function () {
-  const dummyConnectionPool = {
-    getConnection: () => null
-  }
   it('extends Transport', function () {
     assert(AmazonTransport.prototype instanceof Transport)
   })
 
-  it('calls the callback if provided', function (done) {
-    new AmazonTransport({
-      connectionPool: dummyConnectionPool
-    }).request({}, {}, () => done())
-  })
+  describe('.request()', function () {
+    const transport = new AmazonTransport({
+      connectionPool: {
+        getConnection: () => null
+      }
+    })
 
-  it('returns a promise if callback not provided', function (done) {
-    new AmazonTransport({
-      connectionPool: dummyConnectionPool
-    }).request({}, {}).catch(() => done())
-  })
+    it('calls the callback if provided', function (done) {
+      transport.request({}, {}, () => done())
+    })
 
-  it('accepts callback in place of options', function (done) {
-    new AmazonTransport({
-      connectionPool: dummyConnectionPool
-    }).request({}, () => done())
+    it('returns a Promise if callback not provided', function (done) {
+      transport.request({}, {}).catch(() => done())
+    })
+
+    it('accepts callback in place of options', function (done) {
+      transport.request({}, () => done())
+    })
   })
 })
